requirements.txt
langchain
langchain-groq
langchain-core
pydantic
python-dotenv
tenacity
rapidfuzz
